+++ frontend/src/components/Onboarding/Covie.js
